Move auth flow to actions
