--- src/pages/api/auth/signout.ts
@@ -1,8 +1,0 @@
-import type { APIRoute } from "astro";
-
-export const GET: APIRoute = async ({ redirect, cookies }) => {
-  cookies.delete("__session", {
-    path: "/",
-  });
-  return redirect("/signin");
-};
